pyproject.toml
[project]
name = "mcp-for-beginners"
version = "0.1.0"
description = "Model Context Protocol demonstration package using FastMCP"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
authors = [
    {name = "mcp-for-beginners"},
]
keywords = ["mcp", "model-context-protocol", "fastmcp", "demo", "tutorial"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "aiohttp>=3.12.15",
    "azure-ai-inference>=1.0.0b9",
    "fastapi>=0.116.1",
    "fastmcp>=2.11.3",
    "httpx>=0.28.1",
    "mcp>=1.13.1",
    "numpy>=2.3.3",
    "openai>=1.102.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/oscargbocanegra/mcp-for-beginners"
Repository = "https://github.com/oscargbocanegra/mcp-for-beginners.git"
Issues = "https://github.com/oscargbocanegra/mcp-for-beginners/issues"

[project.scripts]
mcp-server = "mcp_demo.server:main"
mcp-client = "mcp_demo.client:main"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["mcp_demo"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.hatch.build.targets.wheel]
packages = ["mcp_demo"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
